ne-mac-ecs: Small updates before adding chunks to the Pool Allocator

+++ NeonEngine/NeonEngine/Core/Memory/Helpers.h
@@ -3,7 +3,7 @@
 #include "Core/Core.h"
 
 namespace Neon { namespace Memory {
-	static size_t GetAlign(size_t n) {
+	static uint8_t GetAlign(size_t n) {
 		return (n + sizeof(intptr_t) - 1) & ~(sizeof(intptr_t) - 1);
 	}
